services/python-staticjs-furever-home/python-backend/app/requirements.txt
requests==2.24.*
mysql-connector-python==8.0.*
Flask==1.1.*
flask-cors==3.0.*
flask-restful